be/feat : modify guide id type

See merge request s09-final/S09P31C103!163

+++ back-end/src/main/java/com/ssafy/eoullim/repository/jpa/GuideRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.eoullim.repository.jpa;
 import com.ssafy.eoullim.model.entity.GuideEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GuideRepository extends JpaRepository<GuideEntity, Long> {
+public interface GuideRepository extends JpaRepository<GuideEntity, Integer> {
 }
